back_end/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.7.1
distlib==0.3.9
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==0.61.1
django-filter==24.3
django-nextjs==3.2.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
filelock==3.16.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
oauthlib==3.2.2
platformdirs==4.3.6
propcache==0.2.0
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
yarl==1.16.0
